Create README.md
An Agricultural Economist working with the National Agricultural Research Organization (NARO) in Uganda and a PhD student of Agricultural and Applied Economics. My work centres on multidisciplinary research with a focus on methods of financial, economic, social, policy, environmental and geographical to contribute to efforts to transform food systems and reduce poverty. Currently, my work involves the integration of the large-scale technology/variety evaluation at the farm level using the Triadic Comparison of Technologies (TRICOT). I embrace open-data sources and the use of Big data. I am sharing the work files in Github.